requirements.txt
altgraph==0.17
astroid==2.4.2
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.3
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
decorator==4.4.2
future==0.18.2
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
itemadapter==0.1.0
jedi==0.17.1
jupyter-client==6.1.5
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
lxml==4.5.1
macholib==1.14
mccabe==0.6.1
numpy==1.19.0
pandas==1.0.5
parsel==1.6.0
parso==0.7.0
pefile==2019.4.18
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
Protego==0.1.16
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pyinstaller==4.0.dev0
pylint==2.5.3
pyOpenSSL==19.1.0
PySide2==5.15.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
queuelib==1.5.0
requests==2.24.0
Scrapy==2.2.0
service-identity==18.1.0
shiboken2==5.15.0
six==1.15.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.22.0
wcwidth==0.2.5
wrapt==1.12.1
zope.interface==5.1.0
